+++ src/cbor/gap.ts
@@ -1,7 +1,15 @@
 // Why is the default value being stored in an array? undefined, null, false, etc... are all valid defaults,
 // and specifying a field on a class as optional will make it undefined by default.
 
+/**
+ * Represents a Gap and its intended value.
+ */
 export type Fill<T = unknown> = [Gap<T>, T];
+
+/**
+ * A Gap represents a placeholder value that can be filled
+ * at a later point in time.
+ */
 export class Gap<T = unknown> {
 	readonly args: [T?] = [];
 	constructor(...args: [T?]) {
