+++ src/index.ts
@@ -1,6 +1,6 @@
 export { Emitter, type Listener, type UnknownEvents } from "./util/emitter.ts";
-export { surql, surrealql } from "./util/taggedTemplate.ts";
-export { PreparedQuery } from "./util/preparedQuery.ts";
+export { surql, surrealql } from "./util/tagged-template.ts";
+export { PreparedQuery } from "./util/prepared-query.ts";
 
 export * as cbor from "./cbor";
 export * from "./cbor/gap";
@@ -9,10 +9,10 @@ export * from "./data";
 export * from "./errors.ts";
 export * from "./types.ts";
 export * from "./util/jsonify.ts";
-export * from "./util/versionCheck.ts";
-export * from "./util/getIncrementalID.ts";
-export * from "./util/stringPrefixes.ts";
-export * from "./util/toSurrealqlString.ts";
+export * from "./util/version-check.ts";
+export * from "./util/get-incremental-id.ts";
+export * from "./util/string-prefixes.ts";
+export * from "./util/to-surrealql-string.ts";
 
 export {
 	ConnectionStatus,
