+++ src/util/process-auth-vars.ts
