+++ tests/unit/convert-auth.test.ts
